// packages/inkibra/api-base/constants/mime-type.ts
export enum MimeType {
  APPLICATION_JSON = 'application/json',
  APPLICATION_PDF = 'application/pdf',
  APPLICATION_DATA = 'application/data',
  IMAGE_JPEG = 'image/jpeg',
  IMAGE_PNG = 'image/png',
  IMAGE_WEBP = 'image/webp',
  TEXT_JAVASCRIPT = 'text/javascript',
  TEXT_HTML = 'text/html',
  TEXT_XML = 'text/xml',
  TEXT_PLAIN = 'text/plain',
  TEXT_CSS = 'text/css',
  TEXT_EVENT_STREAM = 'text/event-stream',
  AUDIO_MPEG = 'audio/mpeg',
  AUDIO_WAV = 'audio/wav',
  AUDIO_MP4 = 'audio/x-m4a',
  VIDEO_MP4 = 'video/mp4',
  MULTIPART_FORM_DATA = 'multipart/form-data',
}